requirements.txt
yfinance
requests
beautifulsoup4
google-generativeai
schedule
streamlit
